Extract document orchestration into client/documents.py
